Move zoom code into useZoomToExistingComponents
